+++ datastructures/vec.cpp
@@ -5,22 +5,22 @@
 #include "vec.h"
 
 // L2 norm of this vector squared
-double vec::norm_sq() const {
+long double vec::norm_sq() const {
     return this->x * this->x + this->y * this->y;
 }
 
 // L2 norm of this vector
-double vec::norm() const {
+long double vec::norm() const {
     return sqrt(this->norm_sq());
 }
 
 // dot product
-double vec::dot(const vec &v) const {
+long double vec::dot(const vec &v) const {
     return this->x * v.x + this->y * v.y;
 }
 
 // size of the cross product, < 0 if right product, > 0 if left product
-double vec::cross(const vec &v) const {
+long double vec::cross(const vec &v) const {
     return this->x * v.y - this->y * v.x;
 }
 
@@ -39,18 +39,18 @@ vec toVec(const vec &a, const vec &b) {
     return vec(a, b);
 }
 
-double norm_sq(const vec &v) {
+long double norm_sq(const vec &v) {
     return v.norm_sq();
 }
 
-double norm(const vec &v) {
+long double norm(const vec &v) {
     return v.norm();
 }
 
-double cross(const vec &a, const vec &b) {
+long double cross(const vec &a, const vec &b) {
     return a.cross(b);
 }
 
-double dot(const vec &a, const vec &b) {
+long double dot(const vec &a, const vec &b) {
     return a.dot(b);
 }
